Moving password help to a helper class
